Refactor MultiStreamDataSampler and data loaders  (#102)
* Initial commit to enable support for source and target channel
selection. This requires a substantial refactoring, which was anyway
necessary to modularize the code. In particular, the batch construction
for source and target has been moved completely to the Batchifyer, which
finally simplifies the MultiStreamDataSampler to a reasonable size. This
change should also facilitate the re-introduction of masked token
modeling.

* Simplified interface of ctor of MultiStreamDataSampler by passing config
and using it wherever possible.

* Cleaned up interface

* Adding more streams---but they need to properly integrated to ensure all
functionality is supported.

* - Fixed various issues in FESOM dataloader and made it compliant with changes in MultiStreamdataSampler
- ruff-ing

* Initial commit to enable support for source and target channel
selection. This requires a substantial refactoring, which was anyway
necessary to modularize the code. In particular, the batch construction
for source and target has been moved completely to the Batchifyer, which
finally simplifies the MultiStreamDataSampler to a reasonable size. This
change should also facilitate the re-introduction of masked token
modeling.

* Simplified interface of ctor of MultiStreamDataSampler by passing config
and using it wherever possible.

* Cleaned up interface

* Adding more streams---but they need to properly integrated to ensure all
functionality is supported.

* - Fixed various issues in FESOM dataloader and made it compliant with changes in MultiStreamdataSampler
- ruff-ing

* Updates